+++ src/mol-canvas3d/helper/camera-helper.ts
@@ -358,6 +358,7 @@ function createMeshRenderObject(props: AxesProps) {
         ...PD.getDefaultValues(Mesh.Params),
         ...props,
         ignoreLight: true,
+        cellSize: 0,
     });
 }
 
@@ -400,5 +401,6 @@ function createTextRenderObject(props: AxesProps) {
         ...PD.getDefaultValues(Text.Params),
         ...props,
         alpha: props.labelOpacity,
+        cellSize: 0,
     });
 }