+++ src/mol-canvas3d/helper/handle-helper.ts
@@ -75,6 +75,7 @@ export class HandleHelper {
                     const params = {
                         ...props.handle.params,
                         scale: props.handle.params.scale * this.webgl.pixelRatio,
+                        cellSize: 0,
                     };
                     this.renderObject = createHandleRenderObject(params);
                     this.scene.add(this.renderObject);