+++ src/mol-gl/shader-code.ts
@@ -21,6 +21,8 @@ export interface ShaderExtensions {
     readonly fragDepth?: ShaderExtensionsValue
     readonly drawBuffers?: ShaderExtensionsValue
     readonly shaderTextureLod?: ShaderExtensionsValue
+    /** Needed to enable the `gl_DrawID` built-in */
+    readonly multiDraw?: ShaderExtensionsValue
     readonly clipCullDistance?: ShaderExtensionsValue
     readonly conservativeDepth?: ShaderExtensionsValue
 }
@@ -250,6 +252,14 @@ function getGlsl100VertPrefix(extensions: WebGLExtensions, shaderExtensions: Sha
             throw new Error(`required 'GL_EXT_draw_buffers' extension not available`);
         }
     }
+    if (shaderExtensions.multiDraw) {
+        if (extensions.multiDraw) {
+            prefix.push('#extension GL_ANGLE_multi_draw : require');
+            prefix.push('#define enabledMultiDraw');
+        } else if (shaderExtensions.multiDraw === 'required') {
+            throw new Error(`required 'GL_ANGLE_multi_draw' extension not available`);
+        }
+    }
     return prefix.join('\n') + '\n';
 }
 
@@ -314,6 +324,14 @@ function getGlsl300VertPrefix(extensions: WebGLExtensions, shaderExtensions: Sha
     if (shaderExtensions.drawBuffers) {
         if (extensions.drawBuffers) {
             prefix.push('#define requiredDrawBuffers');
+        }
+    }
+    if (shaderExtensions.multiDraw) {
+        if (extensions.multiDraw) {
+            prefix.push('#extension GL_ANGLE_multi_draw : require');
+            prefix.push('#define enabledMultiDraw');
+        } else if (shaderExtensions.multiDraw === 'required') {
+            throw new Error(`required 'GL_ANGLE_multi_draw' extension not available`);
         }
     }
     if (shaderExtensions.clipCullDistance) {